+++ EPS Homebridge.indigoPlugin/Contents/Server Plugin/lib/eps.py
@@ -13,8 +13,9 @@ from update import update
 from ui import ui
 from support import support
 
+
 class eps:
-	VERSION = "2.7.0"
+	VERSION = "2.9"
 	
 	#
 	# Initialize the  class
@@ -88,7 +89,12 @@ class eps:
 			if lib == "actions":
 				self.logger.threaddebug("Loading actions library")
 				from actions import actions
-				self.act = actions(self)
+				self.act = actions (self)
+				
+			if lib == "devices":
+				self.logger.threaddebug("Loading device extensions library")
+				from devices import devices
+				self.devices = devices (self)
 								
 			if lib == "conditions":
 				self.logger.threaddebug("Loading conditions library")
@@ -125,9 +131,25 @@ class eps:
 			del plugin.pluginPrefs["debugMode"]
 				
 		
+	#
+	# Call back a plugin method if it exists
+	#
+	def raiseEvent (self, method, args):
+		retval = None
+		
+		try:
+			if method in dir(self.plugin):
+				func = getattr(self.plugin, method)
 				
-				
-				
+				if len(args) > 0: 
+					retval = func(*args)
+				else:
+					retval = func()
+					
+		except Exception as e:
+			self.logger.error (ext.getException(e))			
+			
+		return retval